Change Vector3 to use float instead of double because we do not need as much

+++ OtterEngine/MathEngine/src/Vector3.cpp
@@ -1,24 +1,24 @@
 #include "Vector3.hpp"
 
-Vector3::Vector3(double x1, double y1, double z1){
+Vector3::Vector3(float x1, float y1, float z1){
   x = x1;
   y = y1;
   z = z1;
 }
 
 Vector3 Vector3::operator+(Vector3 other) {
-  double newX = x + other.x;
-  double newY = y + other.y;
-  double newZ = z + other.z;
+  float newX = x + other.x;
+  float newY = y + other.y;
+  float newZ = z + other.z;
 
   Vector3 newVector(newX, newY, newZ);
   return newVector;
 }
 
 Vector3 Vector3::Add(Vector3 other) {
-  double newX = x + other.x;
-  double newY = y + other.y;
-  double newZ = z + other.z;
+  float newX = x + other.x;
+  float newY = y + other.y;
+  float newZ = z + other.z;
 
   Vector3 newVector(newX, newY, newZ);
   return newVector;
